Refactor : 호버 이벤트 성능 최적화 및 username 있을시 출력
