+++ src/pages/DetailPage/TimeTable/VoteTable/VoteCellContainer.tsx
@@ -1,7 +1,7 @@
 import { cx } from '@emotion/css';
 import styled from '@emotion/styled';
 import { MouseEvent, useCallback, useEffect, useState } from 'react';
-import { VotedUserList } from './VotedUserList';
+import { VoteCell } from './VoteCell';
 import { useDispatch, useSelector } from '@/_redux/hooks';
 import { draggingMyCell, setDragPoint } from '@/_redux/slices/timeTableSlice';
 import { IVotedUser } from '@/api/_types/apiModels';
@@ -25,11 +25,6 @@ export const VoteCellContainer = ({
   const dispatch = useDispatch();
   const { prevMyCells, prevVotedCells } = useSelector((state) => state.cells);
   const [isDragging, setIsDragging] = useState(false);
-  const [hoverPosition, setHoverPosition] = useState<{
-    top: null | number;
-    left: null | number;
-  }>({ top: null, left: null });
-  const [hoverIndex, setHoverIndex] = useState([-1, -1]);
 
   const handleMouseEnterMyVote = useCallback(
     (rowIndex: number, columnIndex: number) => {
@@ -58,27 +53,6 @@ export const VoteCellContainer = ({
     },
     [isDragging],
   );
-
-  const handleMouseEnter = useCallback(
-    (e: MouseEvent, i: number, j: number) => {
-      if (i === hoverIndex[0] && j === hoverIndex[1]) return;
-      const rect = e.currentTarget.getBoundingClientRect();
-      const top = rect.top;
-      const left =
-        rect.left +
-        window.scrollX +
-        (e.target as HTMLTableCellElement).offsetWidth;
-
-      setHoverPosition({ top, left });
-      setHoverIndex([i, j]);
-    },
-    [],
-  );
-
-  const handleMouseLeave = useCallback(() => {
-    setHoverPosition({ top: null, left: null });
-    setHoverIndex([-1, -1]);
-  }, []);
 
   useEffect(() => {
     window.addEventListener('mouseup', handleMouseUp);
@@ -131,31 +105,16 @@ export const VoteCellContainer = ({
                       {i % 2 === 0 ? timeColumn[i] : ' '}
                     </StCell>
                     {row.map(({ votedUser, classList }, j) => (
-                      <StVotedCell
+                      <VoteCell
                         key={self.crypto.randomUUID()}
-                        className={cx(classList)}
-                        onMouseEnter={(e) => handleMouseEnter(e, i, j)}
-                        onMouseLeave={handleMouseLeave}
+                        i={i}
+                        j={j}
+                        classList={classList}
                         dateRowLength={dateRow.length}
                         userNum={votedUser.length > 0 ? votedUser.length : ' '}
-                        percentage={
-                          participants.length > 0
-                            ? votedUser.length / participants.length
-                            : 0
-                        }>
-                        {votedUser.length > 0 &&
-                          hoverIndex[0] === i &&
-                          hoverIndex[1] === j && (
-                            <VotedUserList
-                              userList={votedUser}
-                              style={{
-                                position: 'fixed',
-                                top: hoverPosition.top as number,
-                                left: hoverPosition.left as number,
-                              }}
-                            />
-                          )}
-                      </StVotedCell>
+                        votedUser={votedUser}
+                        participants={participants.length}
+                      />
                     ))}
                   </StBodyRow>
                 ))}
@@ -268,71 +227,3 @@ const StCellTable = styled.table`
 `;
 
 const StTableBody = styled.tbody``;
-
-interface IStVotedCell {
-  userNum: number | string;
-  percentage: number;
-  dateRowLength: number;
-}
-
-const StVotedCell = styled.td<IStVotedCell>`
-  position: relative;
-  padding: 0 1px 1px 0;
-  white-space: pre-wrap;
-
-  &::before {
-    box-sizing: border-box;
-    content: ' ';
-    display: block;
-    font-size: 10px;
-    border-radius: 4px;
-    cursor: pointer;
-
-    padding: 6px 2px 6px
-      ${({ dateRowLength }) => (dateRowLength < 5 ? 120 / dateRowLength : 36)}px;
-
-    &:hover {
-      filter: brightness(120%);
-      box-shadow: 0 0 2px 1px olive;
-    }
-
-    background-color: rgba(
-      34,
-      139,
-      180,
-      ${({ percentage }) => 0.2 + percentage * 0.6}
-    );
-    ${({ percentage }) =>
-      percentage === 0 ? 'background-color: #EEEAE2' : ''};
-  }
-
-  &::after {
-    position: absolute;
-    display: inline-block;
-    right: 2px;
-    bottom: 2px;
-    content: '${({ userNum }) => userNum}';
-    font-size: 10px;
-    text-align: end;
-    color: ${({ theme }) => theme.colors.semiWhite};
-  }
-
-  &.voted-mine {
-    &::before {
-      filter: brightness(102%);
-      box-shadow: 0 0 2px 1px #0f2b46;
-    }
-  }
-
-  &.voting {
-    &::before {
-      box-shadow: 0 0 2px 1px #0f2b46;
-    }
-  }
-
-  &.unvoting {
-    &::before {
-      box-shadow: 0 0 2px 1px transparent;
-    }
-  }
-`;
